+++ example/navigation/screens.tsx
@@ -21,8 +21,6 @@ import SelectScreen from '../screens/Select';
 import RadioScreen from '../screens/Radio';
 import SwitchScreen from '../screens/Switch';
 import HeaderScreen from '../screens/Header';
-import HeaderWithCustomBackgroundScreen from '../screens/HeaderWithCustomBackground';
-import HeaderWithRightIconScreen from '../screens/HeaderWithRightIcon';
 import MessageScreen from '../screens/Message';
 import AlertScreen from '../screens/Alert';
 import ModalScreen from '../screens/Modal';
@@ -97,16 +95,6 @@ export const itemsDrawer = [
     name: 'Header',
     options: { headerShown: false },
     component: HeaderScreen,
-  },
-  {
-    name: 'HeaderWithCustomBackground',
-    options: { headerShown: false },
-    component: HeaderWithCustomBackgroundScreen,
-  },
-  {
-    name: 'HeaderWithRightIcon',
-    options: { headerShown: false },
-    component: HeaderWithRightIconScreen,
   },
   {
     name: 'Alert',