+++ website/content/utils/generateCode.ts
@@ -1,7 +1,15 @@
 import { globalStyles } from './constants';
 
-export default function generateCode(code: string, imports?: string) {
-  return `import { View, StyleSheet${imports ? `, ${imports}` : ''} } from 'react-native';
+export default function generateCode(
+  code: string,
+  _imports?: string | string[]
+) {
+  const imports =
+    typeof _imports === 'string' ? _imports : _imports?.join(', ');
+
+  return `import { View, StyleSheet${
+    imports ? `, ${imports.trim()}` : ''
+  } } from 'react-native';
 ${code}
 
 ${globalStyles}