+++ CHANGELOG.md
@@ -1,10 +1,30 @@
+## 1.0.0-beta.1.4.9
+
+`2022-12-02`
+
+### Website
+- Added Header documentation page.
+
+### Package Native
+- Fixed Header icon left or right visibility.
+- Added Navigation Context for pass api props to Header component.
+- Added statusbarHeight prop to Header component.
+- More updates for Header.
+
+### Example
+- Changed title message in header screen.
+- Changed props to header using to Navigation context api in Screen.
+
 ## 1.0.0-beta.1.4.8
 
 `2022-12-01`
 
+### Website
 - Added Carousel documentation page.
 - Added Collapse documentation page.
 - Added Image documentation page.
+
+### Package Native
 - Changed width context from windows to screen.
 - Fixed openKeys prop in collapse.
 
@@ -12,5 +32,6 @@
 
 `2022-12-01`
 
+### Website
 - Added table props for Badge page.
 - Added Card documentation page.